+++ lib/utils/game_options.dart
@@ -3,19 +3,23 @@ import 'package:shared_preferences/shared_preferences.dart';
 class GameOptions {
   static const String _encounterWarningsKey = 'encounterWarnings';
   static const String _mouseInputKey = 'mouseInput';
+  static const String _interfacePgUpKey = 'interfacePgUp';
   bool encounterWarnings = false;
   bool mouseInput = true;
+  String interfacePgUp = "[";
 
   Future<void> load() async {
     final prefs = await SharedPreferences.getInstance();
     encounterWarnings = prefs.getBool(_encounterWarningsKey) ?? false;
     mouseInput = prefs.getBool(_mouseInputKey) ?? true;
+    interfacePgUp = prefs.getString(_interfacePgUpKey) ?? "[";
   }
 
   Future<void> save() async {
     final prefs = await SharedPreferences.getInstance();
     await prefs.setBool(_encounterWarningsKey, encounterWarnings);
     await prefs.setBool(_mouseInputKey, mouseInput);
+    await prefs.setString(_interfacePgUpKey, interfacePgUp);
   }
 }