+++ lib/utils/interface_options.dart
@@ -1,7 +1,18 @@
 import 'package:collection/collection.dart';
 import 'package:lcs_new_age/engine/engine.dart';
+import 'package:lcs_new_age/utils/game_options.dart';
 
-String interfacePgUp = "[";
+String get interfacePgUp => gameOptions.interfacePgUp;
+String get interfacePgDown {
+  if (interfacePgUp == "[") {
+    return "]";
+  } else if (interfacePgUp == ";") {
+    return ".";
+  } else if (interfacePgUp == ",") {
+    return ".";
+  }
+  return "PGDN";
+}
 
 String aOrAn(String word) {
   switch (word[0].toLowerCase()) {
@@ -17,31 +28,21 @@ String aOrAn(String word) {
 }
 
 bool isPageUp(int c) {
-  if (interfacePgUp == "[") {
-    return c == "[".codePoint;
-  } else if (interfacePgUp == ".") {
-    return c == ":".codePoint;
-  } else {
-    return false; // c == "PGUP";
-  }
+  return c == "[".codePoint || c == ";".codePoint;
 }
 
 bool isPageDown(int c) {
-  if (interfacePgUp == "[") {
-    return c == "]".codePoint;
-  } else if (interfacePgUp == ".") {
-    return c == ";".codePoint;
-  } else {
-    return false; // c == "PGDN";
-  }
+  return c == "]".codePoint || c == ".".codePoint;
 }
 
 String get previousPageStr {
   String str;
   if (interfacePgUp == "[") {
     str = "[";
-  } else if (interfacePgUp == ".") {
+  } else if (interfacePgUp == ";") {
     str = ";";
+  } else if (interfacePgUp == ",") {
+    str = ",";
   } else {
     str = "PGUP";
   }
@@ -52,8 +53,8 @@ String get nextPageStr {
   String str;
   if (interfacePgUp == "[") {
     str = "]";
-  } else if (interfacePgUp == ".") {
-    str = ":";
+  } else if (interfacePgUp == ";" || interfacePgUp == ",") {
+    str = ".";
   } else {
     str = "PGDN";
   }
@@ -64,8 +65,10 @@ String get pageStr {
   String str;
   if (interfacePgUp == "[") {
     str = "[]";
-  } else if (interfacePgUp == ".") {
-    str = ";:";
+  } else if (interfacePgUp == ";") {
+    str = ";.";
+  } else if (interfacePgUp == ",") {
+    str = ",.";
   } else {
     str = "PGUP/PGDN";
   }
@@ -76,8 +79,10 @@ String pageStrWithCurrentAndMax(int current, int max) {
   String str;
   if (interfacePgUp == "[") {
     str = "[]";
-  } else if (interfacePgUp == ".") {
-    str = ";:";
+  } else if (interfacePgUp == ";") {
+    str = ";.";
+  } else if (interfacePgUp == ",") {
+    str = ",.";
   } else {
     str = "PGUP/PGDN";
   }
@@ -89,14 +94,29 @@ void addPageButtons(
   y ??= console.y;
   x ??= console.x;
   move(y, x);
-  if (short) {
-    addInlineOptionText("[", "[ - Prev");
-    console.x += 2;
-    addInlineOptionText("]", "] - Next");
+  String pageUpStr;
+  String pageDownStr;
+  if (interfacePgUp == "[") {
+    pageUpStr = "[";
+    pageDownStr = "]";
+  } else if (interfacePgUp == ";") {
+    pageUpStr = ";";
+    pageDownStr = ".";
+  } else if (interfacePgUp == ",") {
+    pageUpStr = ",";
+    pageDownStr = ".";
   } else {
-    addInlineOptionText("[", "[ - Previous Page");
+    pageUpStr = "PGUP";
+    pageDownStr = "PGDN";
+  }
+  if (short) {
+    addInlineOptionText(pageUpStr, "$pageUpStr - Prev");
     console.x += 2;
-    addInlineOptionText("]", "] - Next Page");
+    addInlineOptionText(pageDownStr, "$pageDownStr - Next");
+  } else {
+    addInlineOptionText(pageUpStr, "$pageUpStr - Previous Page");
+    console.x += 2;
+    addInlineOptionText(pageDownStr, "$pageDownStr - Next Page");
   }
   if (current != null && max != null) {
     console.x += 1;